# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bluebulb Expense Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /expenses:
    get:
      operationId: getExpenses
      tags: [expenses]
      summary: Get expense requests
      parameters:
        - name: requesterId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: approverId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseRequest"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create expense request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseRequest"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseRequest"
  /expenses/{id}:
    get:
      operationId: getExpenseById
      tags: [expenses]
      summary: Get expense by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseRequest"
    put:
      operationId: updateExpense
      tags: [expenses]
      summary: Update expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseRequest"
      responses:
        "200":
          description: Expense updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseRequest"
  /expenses/{id}/approve:
    post:
      operationId: approveExpense
      tags: [approvals]
      summary: Approve or reject an expense at current step
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalActionRequest"
      responses:
        "200":
          description: Approval action taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseRequest"
  /expenses/{id}/pay:
    post:
      operationId: markExpensePaid
      tags: [approvals]
      summary: Mark expense as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkPaidRequest"
      responses:
        "200":
          description: Expense marked paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseRequest"
  /retirements:
    get:
      operationId: getRetirements
      tags: [retirements]
      summary: Get retirement requests
      parameters:
        - name: requesterId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of retirements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RetirementRequest"
    post:
      operationId: createRetirement
      tags: [retirements]
      summary: Create retirement request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRetirementRequest"
      responses:
        "201":
          description: Retirement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetirementRequest"
  /retirements/{id}:
    get:
      operationId: getRetirementById
      tags: [retirements]
      summary: Get retirement by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Retirement found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetirementRequest"
  /retirements/{id}/approve:
    post:
      operationId: approveRetirement
      tags: [retirements]
      summary: Approve or reject retirement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalActionRequest"
      responses:
        "200":
          description: Retirement action taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetirementRequest"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, requester, approver_manager, internal_control, finance_manager, finance_team]
        department:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - department
        - createdAt
...
    ExpenseRequest:
      type: object
      properties:
        id:
          type: integer
        expenseName:
          type: string
        expenseType:
          type: string
        purpose:
          type: string
        otherInfo:
          type: string
        department:
          type: string
        requesterId:
          type: integer
        requesterName:
          type: string
        approverId:
          type: integer
        approverName:
          type: string
        internalControlId:
          type: integer
        internalControlName:
          type: string
        financeManagerId:
          type: integer
        financeManagerName:
          type: string
        requestDate:
          type: string
        dueDate:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
        accountName:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [draft, pending_manager, pending_internal_control, pending_finance_manager, pending_payment, paid, rejected]
        currentApprovalStep:
          type: string
        attachmentUrl:
          type: string
        receiptUrl:
          type: string
        managerApprovalDate:
          type: string
        managerRejectionReason:
          type: string
        internalControlApprovalDate:
          type: string
        internalControlRejectionReason:
          type: string
        financeManagerApprovalDate:
          type: string
        financeManagerRejectionReason:
          type: string
        paidDate:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - expenseName
        - expenseType
        - purpose
        - department
        - requesterId
        - requesterName
        - approverId
        - approverName
        - requestDate
        - bankName
        - accountNumber
        - accountName
        - amount
        - status
        - createdAt
        - updatedAt
    CreateExpenseRequest:
      type: object
      properties:
        expenseName:
          type: string
        expenseType:
          type: string
        purpose:
          type: string
        otherInfo:
          type: string
        department:
          type: string
        requesterId:
          type: integer
        approverId:
          type: integer
        internalControlId:
          type: integer
        financeManagerId:
          type: integer
        requestDate:
          type: string
        dueDate:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
        accountName:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [draft, pending_manager]
        attachmentUrl:
          type: string
      required:
        - expenseName
        - expenseType
        - purpose
        - department
        - requesterId
        - approverId
        - requestDate
        - bankName
        - accountNumber
        - accountName
        - amount
        - status
    UpdateExpenseRequest:
      type: object
      properties:
        expenseName:
          type: string
        expenseType:
          type: string
        purpose:
          type: string
        otherInfo:
          type: string
        department:
          type: string
        approverId:
          type: integer
        internalControlId:
          type: integer
        financeManagerId:
          type: integer
        requestDate:
          type: string
        dueDate:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
        accountName:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [draft, pending_manager]
        attachmentUrl:
          type: string
    ApprovalActionRequest:
      type: object
      properties:
        action:
          type: string
          enum: [approve, reject]
        reason:
          type: string
        approverId:
          type: integer
      required:
        - action
        - approverId
    MarkPaidRequest:
      type: object
      properties:
        receiptUrl:
          type: string
        paidById:
          type: integer
      required:
        - paidById
    RetirementRequest:
      type: object
      properties:
        id:
          type: integer
        expenseId:
          type: integer
        expenseName:
          type: string
        requesterId:
          type: integer
        requesterName:
          type: string
        department:
          type: string
        amount:
          type: number
        actualAmount:
          type: number
        purpose:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        approverId:
          type: integer
        approverName:
          type: string
        approvalDate:
          type: string
        rejectionReason:
          type: string
        attachmentUrl:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - expenseId
        - expenseName
        - requesterId
        - requesterName
        - department
        - amount
        - purpose
        - status
        - createdAt
        - updatedAt
    CreateRetirementRequest:
      type: object
      properties:
        expenseId:
          type: integer
        requesterId:
          type: integer
        actualAmount:
          type: number
        purpose:
          type: string
        approverId:
          type: integer
        attachmentUrl:
          type: string
      required:
        - expenseId
        - requesterId
        - actualAmount
        - purpose
        - approverId
...