# lib/api-zod/src/generated/types/expenseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bluebulb Expense Management API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseRequestStatus } from "./expenseRequestStatus";

export interface ExpenseRequest {
  id: number;
  expenseName: string;
  expenseType: string;
  purpose: string;
  otherInfo?: string;
  department: string;
  requesterId: number;
  requesterName: string;
  approverId: number;
  approverName: string;
  internalControlId?: number;
  internalControlName?: string;
  financeManagerId?: number;
  financeManagerName?: string;
  requestDate: string;
  dueDate?: string;
  bankName: string;
  accountNumber: string;
  accountName: string;
  amount: number;
  status: ExpenseRequestStatus;
  currentApprovalStep?: string;
  attachmentUrl?: string;
  receiptUrl?: string;
  managerApprovalDate?: string;
  managerRejectionReason?: string;
  internalControlApprovalDate?: string;
  internalControlRejectionReason?: string;
  financeManagerApprovalDate?: string;
  financeManagerRejectionReason?: string;
  paidDate?: string;
  createdAt: string;
  updatedAt: string;
}