# lib/api-zod/src/generated/types/markPaidRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bluebulb Expense Management API
 * OpenAPI spec version: 0.1.0
 */

export interface MarkPaidRequest {
  receiptUrl?: string;
  paidById: number;
}