# lib/api-zod/src/generated/types/updateExpenseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bluebulb Expense Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateExpenseRequestStatus } from "./updateExpenseRequestStatus";

export interface UpdateExpenseRequest {
  expenseName?: string;
  expenseType?: string;
  purpose?: string;
  otherInfo?: string;
  department?: string;
  approverId?: number;
  internalControlId?: number;
  financeManagerId?: number;
  requestDate?: string;
  dueDate?: string;
  bankName?: string;
  accountNumber?: string;
  accountName?: string;
  amount?: number;
  status?: UpdateExpenseRequestStatus;
  attachmentUrl?: string;
}